requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
httpx==0.27.2
Pillow==10.4.0
python-dotenv==1.0.1
boto3==1.35.0

